renamed the script to acquire images by species. Modified to avoid hardcoding.
